chore(openshift): extract OpenShiftNetworkModule for auth manager
