+++ src/main/java/io/cryostat/platform/internal/OpenShiftPlatformStrategy.java
@@ -44,7 +44,7 @@ import io.cryostat.core.log.Logger;
 import io.cryostat.core.net.JFRConnectionToolkit;
 import io.cryostat.core.sys.FileSystem;
 import io.cryostat.net.AuthManager;
-import io.cryostat.net.OpenShiftAuthManager;
+import io.cryostat.net.openshift.OpenShiftAuthManager;
 
 import dagger.Lazy;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
